Simplify Redis tests
Prevous tests were dependent on a lot of Redis client logic, which is not the goal
of this package. This package provides a container with a running Redis version, how
you interact with it is up to the user. Changed to simpler test with only basic
client usage to make sure container is started and ready to accept connections.