Add test for the "init attributes"
Add test that tests that attributes can be set through the __init__
function and later read through class attr.

This test does not test that the attributes has an actual effect, in the
future it might be of interest to test the integration to the docker
container and ensure that class attributes are propagated properly.