fix: pass doctests, s/doctest/doctests/, run them in gha, s/asyncpg/psycopg/ in doctest, fix keycloak flakiness: wait for first user (#505)
Doctests are run as interpreted code from docstrings. In order to run
these tests, libraries need to be available for the example code, and
usage of async code either needs to be wrapped in an `asyncio` call or
avoided completely.

This PR fixes up all failing doctests and makes `make doctests` target
run successfully again.
Summary:
- Renames Make target `doctest` to `doctests` to follow naming
convention from `tests` target
- Adds `doctests` step to Github Action workflow runs
- Replaces `asyncpg` example from `index.rst` with `psycopg` to be able
to run as a doctest. Also added `psycopg` as dev dependency (`asyncpg`
was already missing from here)
- Fixes Keycloak doctest by providing expected output, also did the same
for regular test
- Also: Fixed `wait_for_container` method in `Keycloak` module to
actually wait for the first user to be created (in order to be able to
authenticate at all) before returning the started container, if the
command is `dev-start`. This is needed in order to prevent race
conditions in flaky tests and for the sample usage code.