fix: url quote passwords (#549)
**Changes**

Updated DbContainer to fix #547 by using `urllib.parse.quote`. I
referenced sqlalchemy's implementation, but have not imported the
library.

I have chosen to make this behaviour occur at all times (can't opt in /
out), as it is common, if not the standard for these urls.

**Tests**

Since DbContainer can't be tested on its own, I put the tests across
various database containers. I have pasted the below as comment in the
test files for the listed modules:

```python
# This is a feature in the generic DbContainer class
# but it can't be tested on its own
# so is tested in various database modules:
# - mysql / mariadb
# - postgresql
# - sqlserver
# - mongodb
```

Note the discussion recommended me to test with oracle, but I was unable
to spin the container up locally (even with colima), so opted to replace
it with mongodb.

Is there a template for PRs for the core library? I am unable to find
one so have opted the above format. Please let me know if I have missed
anything in this PR. Thanks!