fix(core): add empty _configure to DockerContainer (#556)
I've used DbContainer in the past as a parent to a class where I
redefined `_configure()` and did not redefine `start()`. I'm now using
DockerContainer and wanted to follow the same pattern but it doesn't
have a `_configure()` function called in `start()`. I figure this gives
the easy option without being invasive?

tests pass

---------

Co-authored-by: David Ankin <daveankin@gmail.com>