chore: Update pyproject and tests to support Mac's/arm local development (#814)
I want to help contribute to testcontainers-python but issues getting
the local environment running. Starting off with fixing the install +
test suite so myself and other people on Mac can get up and running
faster.

Manual testing done:
This branch was tested on a few different Mac machines the and a Linux
VM (still on arm). `make install` and `make tests` checked in all envs,
I also checked docs, build, and cleans on the Linux VM and a M2 mac.


Notes:
* Poetry markers aren't bullet proof, but works for ibm so it's good
enough for now, else we'd need a more complex install and test build
scripts (for dind tests) which I played around with but ended up
removing to keep the PR smaller and less complex.
* Skipping arm + mac specific tests in the test suite, these are testing
either features that Macs don't have or the community module doesn't
actually support macs. For the sake of contributors this shouldn't
matter, the CI will still be testing everything fully.
* Mac skip reasons given: I'm not deeply familiar with these community
packages or the menewsha of mac docker vs other OS dockers. There might
be better explanations for why it needed to be skipped.

Ticket:
https://github.com/testcontainers/testcontainers-python/issues/813

---------

Co-authored-by: David Ankin <daveankin@gmail.com>